+++ src/collective/tiles/advancedstatic/setuphandlers.py
@@ -16,13 +16,6 @@ class HiddenProfiles(object):
 
 def post_install(context):
     """Post install script"""
-    portal_quickinstaller = api.portal.get_tool('portal_quickinstaller')
-    if not portal_quickinstaller.isProductInstalled('plone.app.mosaic'):
-         # skip if mosaic isn't installed
-        return
-    mosaic_profile = 'collective.tiles.advancedstatic:mosaic_support'
-    setup_tool = api.portal.get_tool('portal_setup')
-    setup_tool.runImportStepFromProfile(mosaic_profile, 'plone.app.registry')
 
 
 def uninstall(context):